commit 0.56 Added TaskProcess

+++ src/main/java/org/example/kihelp_back/task/dto/TaskDto.java
@@ -2,6 +2,7 @@ package org.example.kihelp_back.task.dto;
 
 import org.example.kihelp_back.argument.model.Argument;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -10,8 +11,8 @@ public record TaskDto(
         String title,
         String description,
         String identifier,
-        Integer price,
-        Double discount,
+        BigDecimal price,
+        BigDecimal discount,
         boolean visible,
         String type,
         Long developerId,
