+++ src/main/java/org/example/kihelp_back/task/dto/TaskUpdateDto.java
@@ -1,11 +1,13 @@
 package org.example.kihelp_back.task.dto;
 
+import java.math.BigDecimal;
+
 public record TaskUpdateDto(
         String title,
         String description,
         String identifier,
-        Integer price,
-        Double discount,
+        BigDecimal price,
+        BigDecimal discount,
         boolean visible,
         String type,
         boolean autoGenerate,
